Remove User and email string permissions methods
Remove the PermissionsService methods that accept a User object or email
String as arguments. These methods will either be impossible to
implement once we migrate to the JWT sessions or will just be awkward to
use. This refactor consolidates the methods on those that accept the
Session object as an argument, but in future the session can be read
from the thread local implementation so this argument can be removed
after the migration has been completed.