Prepare teams handling for JWT migration
Some further work is required to change the Teams model to use a string
rather than an int for the id. This will also require the access mapping
to be updated as well. Also, the collection details should return the
team ids in the array rather than the team names. This final API change
will need to be managed and coordinated with Florence by means of the
JWT feature flag.