Update keyring to accept session rather user object
The keyring was required the user object to be passed in because this
used to be used for the user specific keyring implementation. Now that
the keyring has be centralised we need to use the session instead in
preparation for the move to the JWT session implementation at which
zebedee will no longer have access to the data provided by the current
UserService.