Add transitional sessions service
Adds and enables a transitional sessions service implementation that will
move towards the ThreadLocal implementation used in the JWT session service
while continuing to rely on the old sessions logic underneath. This
implementation also gathers the list of groups for the user in order to
mimic the behaviour of the JWT sessions to enable us to transition
further towards the new implementation prior to the actual migration in
order to decrease the risk. This implementation should be short lived
and is only needed to allow for refactoring and testing of the thread
local session model prior to JWT sessions being enabled in an attempt
to decrease the risk and "big bang" nature of the JWT session migration.