Enable thread local based permissions methods
Migrate over to some of the new permissions methods, specifically the
ones that utilise the groups information passed via the session. This
partial implementation is in order to reduce the "big bang" when we
migrate fully over to the JWT sessions by enabling us to bring some of
this new functionality into play early.