Fix issue with JWT read from Florece Token header
Switch the logic so that the JWT token passed as the Authorization
header is the primary/default header to use (i.e. is checked first). The
`X-Florence-Token` header may still be used by some transitional code in
the interim.

Also fix issue with parsing the JWT token from the `X-Florence-Token`
whereby the `Bearer` prefix wasn't being stripped leading to
authorisation failing.