Fix setting collection viewer teams (JWT enabled)
A piece of technical debt involving a hack in zebedee that maps the team
names back to team IDs (because Florence is sending the names when it
shouldn't be) is causing creation of collections to fail with a 500 then
409. This is because the logic used to map the name to the ID is no
longer possible with the JWT sessions enabled and group management moved
to the dp-identity-api. Since Florence actually as the team ID to hand
the cleaner solution is to simply have florence provide the team IDs
rather than names in the first place. This change lays the ground work
for that change.