Gracefully handle invalid team when creating collection
There is a potential for a team to be deleted between the time when the
publishing officer loads the collections screen and the time they submit
creation or updating of a collections preview teams list. In this case
the creation/update of the collection will fail in a potentially
partially modified state.

As suggested in PR, this change will log the issue, but allow the action
to complete so as not to create a situation whereby the collection state
becomes corrupted.