Ensure PDF is generated for both languages
There was an issue when the Welsh PDF was generated but not the
English version. This was because the data_cy.json was picked up
before the data.json resulting in the Set of ContentDetail having
the Welsh version instead of the English (as the code expected).

We now always check if a json for the other language exists
(independently of the language) and generate its PDF if so